+++ tests/controllers/test_copy_nodes_controller.py
@@ -6,7 +6,7 @@ import pytest
 
 from src.main import init_settings
 from src.utils import AppSettings
-from src.script_editor import nuke_se_controllers as se
+from src.script_editor import nuke_controllers as se
 
 pytestmark = pytest.mark.controllers
